% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{apr_fit}
\alias{apr_fit}
\title{Generic No documentation}
\usage{
apr_fit(recon_data, reg_eq, regmethod)
}
\arguments{
\item{data}{A dataframe}

\item{write_folder}{Folder in which to save results}

\item{write_file}{Name of file to be saved}
}
\value{
p Plot
}
\description{
This function needs documentation.
}
