% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{apr_reconst}
\alias{apr_reconst}
\title{Generic No documentation}
\usage{
apr_reconst(recon_model)
}
\arguments{
\item{data}{A dataframe}

\item{write_folder}{Folder in which to save results}

\item{write_file}{Name of file to be saved}
}
\value{
p Plot
}
\description{
This function needs documentation.
}
