% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleo_objects.R
\name{coef.paleo.fit}
\alias{coef.paleo.fit}
\title{Generic No documentation}
\usage{
\method{coef}{paleo.fit}(fit, ...)
}
\arguments{
\item{data}{A dataframe}

\item{write_folder}{Folder in which to save results}

\item{write_file}{Name of file to be saved}
}
\value{
p Plot
}
\description{
This function needs documentation.
}
