% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{fit_dist_generic}
\alias{fit_dist_generic}
\title{Distribution Fit (generic function)}
\usage{
fit_dist_generic(flows, distr, save_prefix, save_plots = TRUE, save_dir)
}
\arguments{
\item{flows}{Vector of numeric flows to be fit.}

\item{distr}{Character name of the probability distribution function}

\item{save_plots}{TRUE or FALSE Plots should be saved to disk? Defaults to TRUE}

\item{save_dir}{Character for save directory. Defaults to working directory}
}
\value{
monthly_ts a dataframe with results of the null model reconstruction
}
\description{
This function is called by fit_norm_dist.  It fits a distribution to a time series and runs several tests to verify the goodness of fit.
}
\seealso{
\code{\link{flow_perc_fit}}
}
\author{
James Stagge, \email{james.stagge@usu.edu}
}
\keyword{normalization,}
\keyword{percentile}
