% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{fit_model}
\alias{fit_model}
\title{Generic No documentation}
\usage{
fit_model(method, regmethod = "lm", reconst_data, annual_norm = NULL,
  monthly_norm = NULL, pred_ts = NULL, reg_eq = NULL,
  monthly_obs = NULL, ...)
}
\arguments{
\item{data}{A dataframe}

\item{write_folder}{Folder in which to save results}

\item{write_file}{Name of file to be saved}
}
\value{
p Plot
}
\description{
This function needs documentation.
}
