% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{fit_norm_dist}
\alias{fit_norm_dist}
\title{Fit Parametric Distribution and test}
\usage{
fit_norm_dist(flow_series, distr, ref_period = NULL, save_prefix = "",
  save_plots = TRUE, save_dir = getwd())
}
\arguments{
\item{distr}{Character name of the probability distribution function}

\item{ref_period}{Vector of 2 years to subset the data. Defaults to NULL (full datasets)}

\item{save_prefix}{Character prefix for saved files. Typically a site name or location.}

\item{save_plots}{TRUE or FALSE Plots should be saved to disk? Defaults to TRUE}

\item{save_dir}{Character for save directory. Defaults to working directory}

\item{flow_data}{Dataframe with columns "year" and "flow". Column "month" should also be included if running monthly.}

\item{time_scale}{Character of either "monthly" or "annual".}
}
\value{
monthly_ts a dataframe with results of the null model reconstruction
}
\description{
This function fits a distribution to a time series of monthly or annual flows for normalization purposes. It alsos run several tests to verify the goodness of fit, using the function fit_perc.
}
\seealso{
\code{\link{fit_perc}}
}
\author{
James Stagge, \email{james.stagge@usu.edu}
}
\keyword{normalization,}
\keyword{percentile}
