% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{flow_reconstr}
\alias{flow_reconstr}
\title{Generic No documentation}
\usage{
flow_reconstr(recon_model, post_proc = FALSE, interpolate = FALSE)
}
\arguments{
\item{data}{A dataframe}

\item{write_folder}{Folder in which to save results}

\item{write_file}{Name of file to be saved}
}
\value{
p Plot
}
\description{
This function needs documentation.
}
