% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{mf_fit}
\alias{mf_fit}
\title{Generic No documentation}
\usage{
mf_fit(obs_ts, wy_first_month = 1)
}
\arguments{
\item{data}{A dataframe}

\item{write_folder}{Folder in which to save results}

\item{write_file}{Name of file to be saved}
}
\value{
p Plot
}
\description{
This function needs documentation.
}
