% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleo_objects.R
\name{paleo.fit}
\alias{paleo.fit}
\title{Generic No documentation}
\usage{
paleo.fit(method, reconst_data, annual_norm = NULL, monthly_norm = NULL,
  first_month_wy = 10, reg_eq = NULL, pred_ts = NULL)
}
\arguments{
\item{data}{A dataframe}

\item{write_folder}{Folder in which to save results}

\item{write_file}{Name of file to be saved}
}
\value{
p Plot
}
\description{
This function needs documentation.
}
