% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleo_objects.R
\name{paleo.norm}
\alias{paleo.norm}
\title{Generic No documentation}
\usage{
paleo.norm(fit, distr, gof, prefix, ref_period, time_scale)
}
\arguments{
\item{data}{A dataframe}

\item{write_folder}{Folder in which to save results}

\item{write_file}{Name of file to be saved}
}
\value{
p Plot
}
\description{
This function needs documentation.
}
