% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleo_objects.R
\name{paleo.ts}
\alias{paleo.ts}
\title{Paleo Time Series Constructor}
\usage{
paleo.ts(ts, time_scale, site_prefix = "", wy_first_month = NULL, ...)
}
\arguments{
\item{data}{A dataframe}

\item{write_folder}{Folder in which to save results}

\item{write_file}{Name of file to be saved}
}
\value{
p Plot
}
\description{
This function provides a standard time series plot using ggplot's geom_line.
}
