% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_indices.R
\name{shift_indices}
\alias{shift_indices}
\title{Shift Indices}
\usage{
shift_indices(data, name, lags)
}
\arguments{
\item{annual_rec}{a dataframe with columns "water_year" and "annual_flow"}

\item{monthly_prop}{a dataframe with columns "month" and "prop"}

\item{first_month_wy}{a numeric variable with the month that signifies the start of the water year, usually 10}
}
\value{
monthly_ts a dataframe with results of the null model reconstruction
}
\description{
No description
}
\examples{
mf_model()

}
